test: move tools/parser to tests
